replace all doubles/longs by BigInteger/BigDecimal

+++ src/main/java/com/valdist/exception/processing/AccountHasNotFoundException.java
@@ -1,7 +1,9 @@
 package com.valdist.exception.processing;
 
+import java.math.BigInteger;
+
 public class AccountHasNotFoundException extends ProcessingCenterException {
-    public AccountHasNotFoundException(long accountNumber) {
+    public AccountHasNotFoundException(BigInteger accountNumber) {
         super("Account has not been found by number " + accountNumber);
     }
 }
